# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Biblioteca",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/usuarios/{idUsuario}/libros": {
            "get": {
                "tags": [
                    "Prestamos"
                ],
                "summary": "Mostrar todos los libros que tiene determinado usuario",
                "operationId": "4524370cd5ce286b9bb64f4c405abe02",
                "parameters": [
                    {
                        "name": "idUsuario",
                        "in": "path",
                        "description": "La id del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar todos los libros de ese usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Prestamos"
                ],
                "summary": "Añadir un libro a determinado usuario",
                "operationId": "fe243cac67ccfc832153b7c183ec5271",
                "parameters": [
                    {
                        "name": "idUsuario",
                        "in": "path",
                        "description": "La id del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "libro_id",
                        "in": "query",
                        "description": "Id del libro prestado a usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios con libros."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/usuarios/{idUsuario}/libros/{idLibro}": {
            "post": {
                "tags": [
                    "Prestamos"
                ],
                "summary": "Añadir un libro a determinado usuario",
                "operationId": "202aa48594d95f0759d4c048b7d8435b",
                "parameters": [
                    {
                        "name": "idUsuario",
                        "in": "path",
                        "description": "La id del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "idLibro",
                        "in": "path",
                        "description": "La id del libro a nodificar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "libro_id",
                        "in": "query",
                        "description": "La Id del nuevo libro para prestado a usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios con libros."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Prestamos"
                ],
                "summary": "Borrar un libro a  un determinado usuario",
                "operationId": "d9c64ac5f256e889e918d7adef6b12dd",
                "parameters": [
                    {
                        "name": "idUsuario",
                        "in": "path",
                        "description": "La id del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "idLibro",
                        "in": "path",
                        "description": "La id del libro a nodificar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios con libros."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        }
    }
}